+++ packages/paraglide/paraglide-next/examples/app/next.config.mjs
@@ -1,0 +1,10 @@
+import { paraglide } from "@inlang/paraglide-next/plugin"
+
+export default paraglide({
+	paraglide: {
+		project: "./project.inlang",
+		outdir: "./src/paraglide",
+	},
+
+	basePath: "/base",
+})